bdp2_oct_2021-group_3-main/requirements.txt
boto3==1.26.89
keras==2.11.0
keras_preprocessing==1.1.2
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
tensorflow==2.11.0
tweepy==4.13.0
imblearn==0.0
